refactor: config to the config module, credentials to the vault
agent_config.c could not move into the config module: module-bus-boundary-check
rejected it for reaching straight into vault/runtime_secret.h,
vault/vault_principal.h and vault/vault_service.h. That was only ever legal
because the file sat in src/server/, outside any module. The name said config; a
third of the contents were credential resolution and thread-local request state.

SPLIT ALONG THE REAL SEAM. Credentials are never config, so they go where
credentials live:

  src/modules/vault/agent_credentials.c  (new)
    agent_expand_env, the provider->env-var tables, agent_provider_env_value,
    agent_vault_get, agent_has_resolvable_credentials, the codex OAuth client +
    refresh + parse block, agent_resolve_auth, agent_build_extra_headers, and the
    per-request identity (session id, vault principal, codex creds, cancel flag).
    None of it is configuration: it answers "can this agent authenticate right
    now?" from the vault, the environment and the live request, all of which
    change without agents.json changing.

  src/modules/config/agent_config.c  (moved)
    parsing, saving, normalising and routing over agents.json. No vault include
    remains, which is what let the move land.

module-bus-boundary: ok (165 declared crossings remain) -- the crossing is gone
rather than declared, because the dependency itself was wrong.

MINIMAL DOCS ON VAULT CONTENTS. gen-reference-docs.py scans the config module
only, deliberately. The generated agents.json table dropped access_token,
refresh_token, exp and tokens -- enumerating the credential-bearing fields, or
naming the file that reads them, documents an attack surface for no operator
benefit. `aimee agent setup` prompts for what an operator actually needs.

Also carried: CMakeLists (which still named the old path, and had never listed
agent_registry.c), the config and vault descriptors, both module pins, and the
four test targets that link the loader and now need its credential half.

make lint: all 55 checks pass. Unit suite: 17 binaries, zero failures.
