+++ dependencies/aimee-repositories.lock.json
@@ -35,7 +35,7 @@
         "server",
         "kb"
       ],
-      "source_sha256": "80adfbc8fce5e871056dbf8cccc571dd2c183dedf5b5d8895a27edc3839f4797"
+      "source_sha256": "d0e4e53f81f8c8f0331729aca3f5e39ea2df2275842a8048339c40228c71a3fe"
     },
     {
       "id": "ir",
@@ -306,7 +306,7 @@
         "server",
         "kb"
       ],
-      "source_sha256": "a5443825a19a5c1c1a2994cf4ddf76640362f72c6b5c999d830a0e791f1aa7c5"
+      "source_sha256": "c69e17b3d00ee4c31a39404187192bac458eeea78efaf5c63e4d323cae2d1bd7"
     },
     {
       "id": "execution-policy",
