+++ tests/baselines/refactor/index.json
@@ -27,7 +27,7 @@
       "files": [
         {
           "path": "docs/gen/configuration.md",
-          "sha256": "786127cb796d5fff1eb24fb999a88cd15686ea7bc11027bdf7b9ca5a121e9512"
+          "sha256": "bbf2da600bbadba8d14b29a939a91aeef1c60c16a4a8da558c76fd488506526b"
         }
       ]
     },
